Notifications/email changes
Updated the email notifications to send emails to ARCH for all 3 users when a case is approved or rejected (with approval letter attached as a pdf), and also to insert notifications into the db so that it can be viewed in the internal notifications center. Also, I modified the notifications code based on the suggested changes.